refactor trace collector span tests
